// src/store/mutation-types.ts
export enum MutationType {
  RESET_DATA = 'RESET_DATA',
  SET_CURRENT_USER = 'SET_CURRENT_USER',
  SET_USER_PROFILE = 'SET_USER_PROFILE',

  ADD_PROJECT = 'ADD_PROJECT',
  EDIT_PROJECT = 'EDIT_PROJECT',
  REMOVE_PROJECT = 'REMOVE_PROJECT',
  REMOVE_ALL_PROJECTS = 'REMOVE_ALL_PROJECTS',

  ADD_FEATURE = 'ADD_FEATURE',
  EDIT_FEATURE = 'EDIT_FEATURE',
  REMOVE_FEATURE = 'REMOVE_FEATURE',
  REMOVE_ALL_FEATURES = 'REMOVE_ALL_FEATURES',

  ADD_ITEM = 'ADD_ITEM',
  EDIT_ITEM = 'EDIT_ITEM',
  REMOVE_ITEM = 'REMOVE_ITEM',
  REMOVE_ALL_ITEMS = 'REMOVE_ALL_ITEMS',

  ADD_COMMENT = 'ADD_COMMENT',
  EDIT_COMMENT = 'EDIT_COMMENT',
  REMOVE_COMMENT = 'REMOVE_COMMENT',
  REMOVE_ALL_COMMENTS = 'REMOVE_ALL_COMMENTS',

  ADD_LINK = 'ADD_LINK',
  EDIT_LINK = 'EDIT_LINK',
  REMOVE_LINK = 'REMOVE_LINK',
  REMOVE_ALL_LINKS = 'REMOVE_ALL_LINKS',
}